requirements.txt
# Core Dependencies
pandas>=2.0.0,<3.0.0
requests>=2.28.0,<3.0.0
pyyaml>=6.0,<7.0
python-dotenv>=1.0.0,<2.0.0
numpy>=1.21.0,<2.0.0

# Development Dependencies
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
black>=23.0.0,<24.0.0
mypy>=1.0.0,<2.0.0
flake8>=6.0.0,<7.0.0
